Keep both 'Hard fork activation' (this branch) and 'Authenticated withdrawals' (main) sections.

Amp-Thread-ID: https://ampcode.com/threads/T-019d214f-8bec-7112-ae30-cb172c4f3379

+++ docs/specs/test/zone/ZonePortal.t.sol
@@ -134,7 +134,7 @@ contract ZonePortalTest is BaseTest {
     GasConsumingReceiver public gasConsumingReceiver;
     SuccessfulReceiver public successfulReceiver;
 
-    uint64 public testZoneId;
+    uint32 public testZoneId;
     bytes32 public constant GENESIS_BLOCK_HASH = keccak256("genesis");
     bytes32 public constant GENESIS_TEMPO_BLOCK_HASH = keccak256("tempoGenesis");
     uint64 public genesisTempoBlockNumber;
